+++ backend/app/api/routes/event.py
@@ -1,13 +1,16 @@
 from typing import List
-from fastapi import APIRouter
+from fastapi import APIRouter, HTTPException, status
 from app.crud import event as crud
 from app.schemas import event as schemas
-from app.api.deps import SessionDep
+from app.api.deps import SessionDep, CurrentUser
+from app.models.userRole import UserRole
 
 router = APIRouter(tags=["Event"])
 
 @router.post("/", response_model=schemas.Event)
-def create_event(db: SessionDep, event: schemas.EventCreate):
+def create_event(db: SessionDep, event: schemas.EventCreate, current_user: CurrentUser):
+    if current_user.role not in [UserRole.ADMIN, UserRole.ORGANIZER]:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not enough permissions")
     return crud.create_event(db=db, event=event)
 
 @router.get("/{event_id}", response_model=schemas.Event)
@@ -19,9 +22,13 @@ def get_events(db: SessionDep):
     return crud.get_events(db=db)
 
 @router.put("/{event_id}", response_model=schemas.Event)
-def update_event(db: SessionDep, event_id: int, event: schemas.EventUpdate):
+def update_event(db: SessionDep, event_id: int, event: schemas.EventUpdate, current_user: CurrentUser):
+    if current_user.role not in [UserRole.ADMIN, UserRole.ORGANIZER]:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not enough permissions")
     return crud.update_event(db=db, event_id=event_id, event=event)
 
 @router.delete("/{event_id}", response_model=schemas.Event)
-def delete_event(db: SessionDep, event_id: int):
+def delete_event(db: SessionDep, event_id: int, current_user: CurrentUser):
+    if current_user.role not in [UserRole.ADMIN, UserRole.ORGANIZER]:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not enough permissions")
     return crud.delete_event(db=db, event_id=event_id)
